// src/generated/model/adresserFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type {AdresserFrontendValg} from "./adresserFrontendValg";
import type {AdresseFrontend} from "./adresseFrontend";
import type {NavEnhetFrontend} from "./navEnhetFrontend";

export interface AdresserFrontend {
    valg?: AdresserFrontendValg;
    folkeregistrert?: AdresseFrontend;
    midlertidig?: AdresseFrontend;
    soknad?: AdresseFrontend;
    navEnhet?: NavEnhetFrontend;
}